requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
config==0.5.1
Django==4.1.1
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-environ==0.10.0
django-filter==23.1
django-htmx==1.14.0
django-tables2==2.5.3
environ==1.0
gunicorn==20.1.0
idna==3.4
mysqlclient==2.1.1
numpy==1.24.2
psycopg2==2.9.4
requests==2.28.2
sqlparse==0.4.2
urllib3==1.26.15
boto3
